flask/requirements.txt
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.2
